requirements.txt
click==6.6
Flask==0.12
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
Pillow==3.4.2
pytesseract==0.1.6
requests==2.12.4
Werkzeug==0.11.15
